% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-trans.R
\name{athena_type_trans}
\alias{athena_type_trans}
\alias{to_cols}
\title{Translate from one type system to another}
\usage{
athena_type_trans(type, to = c("r", "athena"))

to_cols(name, type)
}
\arguments{
\item{type}{type (character)}

\item{to}{one of `athena` or `r`}

\item{name, type}{equal length character vectors; type should be an R type}
}
\description{
Translate from one type system to another
}
