% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsathena-package.R
\docType{package}
\name{awsathena}
\alias{awsathena}
\title{rJava Client to AWS Athena SDK}
\description{
Provides R wrapper methods to core 'aws-java-sdk-athena' Java library methods
to interrogate and orchestrate Amazon Athena queries.
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
