% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-query-ex-res.R
\name{download_query_execution_results}
\alias{download_query_execution_results}
\title{Use S3 to download the results of an Athena Query}
\usage{
download_query_execution_results(
  qxid,
  output_dir,
  progress = FALSE,
  region = "us-east-1",
  profile = NULL,
  buffer_size = 16384L,
  properties_file = NULL
)
}
\arguments{
\item{qxid}{query execution id}

\item{output_dir}{either a full or relative path (\code{\link[=path.expand]{path.expand()}} will be run on
the value) to where you want the results to be stored.}

\item{progress}{show download progress?}

\item{region}{AWS region string}

\item{profile}{if not using the default credentials chain or a dedicated
properties file then provide the named profile from \verb{~/.aws/credentials}
you wish to use}

\item{buffer_size}{S3 temp buffer size; bigger = faster d/l}

\item{properties_file}{if not using the default credentials provider chain or
a named profile then provide the path to an Athena credentials proeprty file.}
}
\description{
This method is much faster than the paginated interface
}
