% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-query-executions.R
\name{list_query_executions}
\alias{list_query_executions}
\title{List Query Executions}
\usage{
list_query_executions(
  region = "us-east-1",
  profile = NULL,
  properties_file = NULL,
  max = NULL
)
}
\arguments{
\item{region}{AWS region string}

\item{profile}{if not using the default credentials chain or a dedicated
properties file then provide the named profile from \verb{~/.aws/credentials}
you wish to use}

\item{properties_file}{if not using the default credentials provider chain or
a named profile then provide the path to an Athena credentials proeprty file.}

\item{max}{maximum number of query execution ids to return. If \code{NULL} only the
most recent query execution ids will be returned. Otherwise sufficient
iterations of a call to the paginated API will be made to retrieve \code{max}
values. Note that the final total number of returned query execution ids
may be more or less than \code{max} depending on the number of paginated
results and no attempt will be made to truncate to \code{max}.}
}
\description{
List Query Executions
}
