% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-download-file.R
\name{s3_download_file}
\alias{s3_download_file}
\title{Download a key from a bucket to a local file}
\usage{
s3_download_file(bucket, key, output_dir, progress = FALSE,
  region = "us-east-1", profile = NULL, properties_file = NULL)
}
\arguments{
\item{bucket, key}{S3 bucket and key (no \code{s3://} prefix)}

\item{output_dir}{where to store \code{key}}

\item{progress}{show download progress?}

\item{region}{AWS region string}

\item{profile}{if not using the default credentials chain or a dedicated
properties file then provide the named profile from \code{~/.aws/credentials}
you wish to use}

\item{properties_file}{if not using the default credentials provider chain or
a named profile then provide the path to an Athena credentials proeprty file.}
}
\description{
Download a key from a bucket to a local file
}
