% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-query-execution.R
\name{start_query_execution}
\alias{start_query_execution}
\title{Start Query Execution}
\usage{
start_query_execution(query, database, output_location,
  client_request_token = uuid::UUIDgenerate(),
  encryption_option = NULL, kms_key = NULL, region = "us-east-1",
  profile = NULL, properties_file = NULL)
}
\arguments{
\item{query}{SQL query statements to be executed}

\item{database}{database within which the query executes.}

\item{output_location}{location in S3 where query results are stored.}

\item{client_request_token}{unique case-sensitive string used to ensure the
request to create the query is idempotent (executes only once). If another
\code{StartQueryExecution} request is received, the same response is returned
and another query is not created. If a parameter has changed, for example,
the \code{query} , an error is returned. \strong{This is auto-generated-for-you}.}

\item{encryption_option}{indicates whether Amazon S3 server-side encryption
with Amazon S3-managed keys (\code{SSE-S3}), server-side encryption with
KMS-managed keys (\code{SSE-KMS}), or client-side encryption with KMS-managed
keys (\code{CSE-KMS}) is used. Default is \code{NULL} (no encryption)}

\item{kms_key}{For \code{SSE-KMS} and \code{CSE-KMS}, this is the KMS key ARN or ID.
Default is \code{NULL} (no encryption)}

\item{region}{AWS region string}

\item{profile}{if not using the default credentials chain or a dedicated
properties file then provide the named profile from \code{~/.aws/credentials}
you wish to use}

\item{properties_file}{if not using the default credentials provider chain or
a named profile then provide the path to an Athena credentials proeprty file.}
}
\description{
Start Query Execution
}
